Move all generated by VS files during building to one folder

+++ src/eevset.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int main(size_t argc, const char* argv[])
+int main(int argc, const char* argv[])
 {
 
     return 0;
